+++ src/colors_utils.c
@@ -4,19 +4,19 @@ t_rgb	rgb_mul_arr(t_rgb *s_rgb1, t_rgb *s_rgb2)
 {
 	t_rgb	s_rgb_mul_arr;
 
-	s_rgb_mul_arr.red = (int)((double)s_rgb1->red * (double)s_rgb2->red * 0.003921568627451);
-	s_rgb_mul_arr.green = (int)((double)s_rgb1->green * (double)s_rgb2->green * 0.003921568627451);
-	s_rgb_mul_arr.blue = (int)((double)s_rgb1->blue * (double)s_rgb2->blue * 0.003921568627451);
+	s_rgb_mul_arr.red = (int)((float)s_rgb1->red * (float)s_rgb2->red * 0.003921568627451);
+	s_rgb_mul_arr.green = (int)((float)s_rgb1->green * (float)s_rgb2->green * 0.003921568627451);
+	s_rgb_mul_arr.blue = (int)((float)s_rgb1->blue * (float)s_rgb2->blue * 0.003921568627451);
 	return (s_rgb_mul_arr);
 }
 
-t_rgb	rgb_mul(t_rgb *s_rgb, double num)
+t_rgb	rgb_mul(t_rgb *s_rgb, float num)
 {
 	t_rgb	s_rgb_mul;
 
-	s_rgb_mul.red = (int)((double)s_rgb->red * num);
-	s_rgb_mul.green = (int)((double)s_rgb->green * num);
-	s_rgb_mul.blue =  (int)((double)s_rgb->blue * num);
+	s_rgb_mul.red = (int)((float)s_rgb->red * num);
+	s_rgb_mul.green = (int)((float)s_rgb->green * num);
+	s_rgb_mul.blue =  (int)((float)s_rgb->blue * num);
 	if (s_rgb_mul.red > 255)
 		s_rgb_mul.red = 255;
 	if (s_rgb_mul.green > 255)
@@ -56,7 +56,7 @@ t_rgb	rgb_average(t_rgb *s_rgb1, t_rgb *s_rgb2, int flag)
 	return (*s_rgb1);
 }
 
-void	rgb_add_light(t_rgb *s_rgb_res, t_rgb *s_rgb_obj, t_rgb *s_rgb_light, double intensity)
+void	rgb_add_light(t_rgb *s_rgb_res, t_rgb *s_rgb_obj, t_rgb *s_rgb_light, float intensity)
 {
 	t_rgb	s_rgb_tmp;
